+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: allefebv <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/09 10:33:56 by allefebv          #+#    #+#             */
-/*   Updated: 2018/11/15 12:34:42 by allefebv         ###   ########.fr       */
+/*   Updated: 2018/11/15 14:13:59 by allefebv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	int		i;
+	size_t	i;
 
 	i = ft_strlen(s);
 	while (i && s[i] != c)
